added final checkout functionality
