+++ src/components/Checkout/billing.module.css
@@ -1,0 +1,3 @@
+input{
+    color: "#9F9F9F"
+}
