+++ src/components/Checkout/checkout.module.css
@@ -2,6 +2,7 @@
     padding: 63px 100px;   
 
 }
+
 @media only screen and (min-width:320px) and (max-width:768px){
 
     .container{
