--- src/trial/form2.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const FormButton = ({ onSubmit }) => {
-  return <button onClick={onSubmit}>Submit</button>;
-};
-
-export default FormButton;
